requirements.txt
argh==0.26.2
astroid==2.12.10
blessed==1.19.1
click-completion==0.5.2
click-default-group==1.2.2
commonmark==0.9.1
cucumber-tag-expressions==4.1.0
dill==0.3.5.1
gitdb==4.0.9
GitPython==3.1.27
greenlet==1.1.3
inquirer==2.10.0
isort==5.10.1
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
pathspec==0.10.1
platformdirs==2.5.2
pluggy==1.0.0
pprintpp==0.4.0
Pygments==2.13.0
python-dotenv==0.21.0
python-editor==1.0.4
PyYAML==6.0
readchar==4.0.3
rich==12.5.1
shellingham==1.5.0
six==1.16.0
smmap==5.0.0
tinydb==4.7.0
tomli==2.0.1
tomlkit==0.11.5
types-PyYAML==6.0.12
ward==0.66.1b0
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
